+++ MapboxNavigation/NavigationViewController.swift
@@ -148,16 +148,29 @@ open class NavigationViewController: UIViewController, NavigationStatusPresenter
     public var shouldManageApplicationIdleTimer = true
     
     /**
-     Enables highlighting of the destination building when arriving.
+     Enum denoting the types of the destination building highlighting on arrival.
      */
-    public var highlightsDestinationBuildings: Bool = false
+    public enum DestinationBuildingHighlightType: Int {
+        /**
+         Do not highlight destination building on arrival.
+         */
+        case noHighlight
+        
+        /**
+         Highlight destination building on arrival in 2D.
+         */
+        case highlightIn2D
+        
+        /**
+         Highlight destination building on arrival in 3D.
+         */
+        case highlightIn3D
+    }
     
     /**
-     When destination building highlighting is enabled, this option controls whether the building will be presented in 3d or 2d.
-     
-     The default value of this property is `true`.
+     Allows to control highlighting of the destination building on arrival. By default destination buildings will not be highlighted.
      */
-    public var highlightBuildingsIn3D: Bool = true
+    public var destinationBuildingHighlightType: DestinationBuildingHighlightType = .noHighlight
     
     var isConnectedToCarPlay: Bool {
         if #available(iOS 12.0, *) {
@@ -593,7 +606,7 @@ extension NavigationViewController: NavigationServiceDelegate {
     }
     
     private func zoomInAndHighlightBuilding(for location: CLLocation?) {
-        if !highlightsDestinationBuildings { return }
+        if destinationBuildingHighlightType == .noHighlight { return }
         guard let mapViewController = self.mapViewController else { return }
         guard let location = location else { return }
         
@@ -613,7 +626,7 @@ extension NavigationViewController: NavigationServiceDelegate {
                           animated: true,
                           completionHandler: {
                             // Highlight buildings which were marked as target destination coordinate in waypoint.
-                            mapView.highlightBuildings(at: self.routeOptions.waypoints.compactMap({ $0.targetCoordinate }), in3D: self.highlightBuildingsIn3D)
+                            mapView.highlightBuildings(at: self.routeOptions.waypoints.compactMap({ $0.targetCoordinate }), in3D: self.destinationBuildingHighlightType == .highlightIn3D ? true : false)
                             
                             // Update insets to be able to correctly center map view after presenting end of route view.
                             mapViewController.updateMapViewContentInsets()